+++ tests/testthat/test-Lagged.R
@@ -1,0 +1,86 @@
+
+test_that("Lagged classes: indexing",
+{
+    v <- 1:12
+    v_lagged <- Lagged(v)
+    expect_identical(v_lagged, new("Lagged1d", data = v))
+    expect_equal(v_lagged[0:2], v[1:3])
+    expect_equal(v_lagged[[0]], 1)
+    expect_equal(v_lagged[0],   v_lagged[0, drop = FALSE])
+    expect_equal(v_lagged[[0]], v_lagged[0, drop = TRUE])
+
+
+    m <- matrix(1:12, nrow = 4)
+    m_lagged <- Lagged(m)
+    expect_identical(m_lagged, new("Lagged2d", data = m))
+    expect_equal(dim(m_lagged[0]), c(4, 1))
+    expect_equal(m_lagged[0], m[ , 1, drop = FALSE])
+
+    expect_null(dim(m_lagged[[0]]))
+    expect_equal(m_lagged[[0]], m[ , 1, drop = TRUE])
+    expect_equal(m_lagged[0],   m_lagged[0, drop = FALSE])
+    expect_equal(m_lagged[[0]], m_lagged[0, drop = TRUE])
+
+
+    a <- array(1:24, dim = c(2, 3, 4))
+    a_lagged <- Lagged(a)
+    expect_identical(a_lagged, new("Lagged3d", data = a))
+    expect_equal(dim(a_lagged[0]), c(2,3,1) ) # c(dim(a)[-3], 1) )
+    expect_equal(a_lagged[0], a[ , , 1, drop = FALSE])
+
+    expect_equal(dim(a_lagged[[0]]), c(2,3) ) # dim(a)[-3]
+    expect_equal(a_lagged[[0]], a[ , , 1, drop = TRUE])
+    expect_equal(a_lagged[0],   a_lagged[0, drop = FALSE])
+    expect_equal(a_lagged[[0]], a_lagged[0, drop = TRUE])
+
+
+    ## as above for "FlexibleLagged"
+    v_flex <- new("FlexibleLagged", data = v)
+    expect_identical(v_flex@data, v_lagged)
+    expect_equal(v_flex[0], v_lagged[0])
+    expect_equal(v_flex[[0]], v_lagged[[0]])
+    expect_equal(v_flex[0],   v_flex[0, drop = FALSE])
+    expect_equal(v_flex[[0]], v_flex[0, drop = TRUE])
+
+    m_flex <- new("FlexibleLagged", data = m)
+    expect_identical(m_flex@data, m_lagged)
+    expect_equal(m_flex[0], m_lagged[0])
+    expect_equal(m_flex[[0]], m_lagged[[0]])
+    expect_equal(m_flex[0],   m_flex[0, drop = FALSE])
+    expect_equal(m_flex[[0]], m_flex[0, drop = TRUE])
+
+    a_flex <- new("FlexibleLagged", data = a)
+    expect_identical(a_flex@data, a_lagged)
+    expect_equal(a_flex[0], a_lagged[0])
+    expect_equal(a_flex[[0]], a_lagged[[0]])
+    expect_equal(a_flex[0],   a_flex[0, drop = FALSE])
+    expect_equal(a_flex[[0]], a_flex[0, drop = TRUE])
+
+    ## maxLag, maxLag<-
+    ##    TODO: extending with "maxLag<-"()
+
+    expect_equal(maxLag(v_lagged), 11)
+    expect_equal(maxLag(m_lagged), 2)
+    expect_equal(maxLag(a_lagged), 3)
+
+    v2_lagged <- v_lagged
+    maxLag(v2_lagged) <- 2
+    expect_equal(maxLag(v2_lagged), 2)
+    expect_equal(v2_lagged[], v[1:3])
+
+    m2_lagged <- m_lagged
+    maxLag(m2_lagged) <- 2
+    expect_equal(maxLag(m2_lagged), 2)
+    expect_equal(m2_lagged[], m[, 1:3])
+
+    a2_lagged <- a_lagged
+    maxLag(a2_lagged) <- 2
+    expect_equal(maxLag(a2_lagged), 2)
+    expect_equal(a2_lagged[], a[, , 1:3])
+
+
+
+})
+
+
+